Added Makefile

+++ src/PointLib.h
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 
 struct Point {
-    int x = 0;
-    int y = 0;
-    int z = 0;
+    float x = 0;
+    float y = 0;
+    float z = 0;
 }
